experiment-20/backend/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
coverage==7.13.5
cryptography==44.0.2
Flask==3.1.2
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
gunicorn==25.1.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
pytest-cov==7.1.0
Werkzeug==3.1.5